+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Client notification mechanism. A task can generate a notification,
   which the process that submitted the task can capture.
 * Server scheduler more respects priorities
+* Server scheduler improved when resource variants are used
 * Extended events: 
   * New journal events: `job-idle`, `task-notify`.
   * `worker-new` now contains the worker configuration,